fix(robot): 调整robot的显示控制
