+++ examples/sites/demos/pc/menus.js
@@ -175,7 +175,6 @@ export const cmpMenus = [
     'key': 'cmp-table-components',
     'children': [
       { 'nameCn': '基本用法', 'name': '', 'key': 'grid' },
-      { 'nameCn': 'AI智能体', 'name': '', 'key': 'grid-ai-agent' },
       { 'nameCn': '序号列', 'name': '', 'key': 'grid-serial-column' },
       { 'nameCn': '选中行', 'name': '', 'key': 'grid-operation-column' },
       { 'nameCn': '空数据', 'name': '', 'key': 'grid-empty' },
