+++ examples/sites/src/router.js
@@ -1,6 +1,7 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import Layout from '@/views/layout/layout.vue'
 import { LANG_PATH_MAP, ZH_CN_LANG, DEFAULT_THEME } from './const'
+import { appData } from './tools/appData.js'
 
 const Components = () => import('@/views/components-doc/index.vue')
 const Docs = () => import('@/views/docs/docs.vue')
@@ -57,5 +58,7 @@ router.afterEach((to, from) => {
   if (to.meta.title) {
     document.title = to.meta.title
   }
+  // tiny-robot 通过路由,确定浮动区，是否显示AI按钮
+  appData.hasFloatRobot = to.path.endsWith('components/grid')
 })
 export { router }
